refactored code to accomodate multiple data in state
